chore(reduce): expose reduce logic
